deleting the func because it didn't work
