+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/res/Hardware.java
@@ -64,34 +64,8 @@ public class Hardware {
         imu.initialize(parameters);
 
     }
-    public static OpenCvCamera camera;
+
     public static void webcamReg() {
-        //initiates the camera
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-        //gets the camera name from the int in the driver station
-        WebcamName webcamName = hardwareMap.get(WebcamName.class, "camera");
-        //to get the webcam view
-        camera = OpenCvCameraFactory.getInstance().createWebcam(webcamName, cameraMonitorViewId);
-        //creates an object to the telemetry from the file rod detector
-        //RodDetector detector = new RodDetector(telemetry);
-        contours contours = new contours(telemetry);
-        //sets the pipeline
-        //camera.setPipeline(detector);
-        camera.setPipeline(contours);
-        //opening the camera
-        camera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
-        {
-            @Override
-            //live streaming
-            public void onOpened()
-            {
-                camera.startStreaming(320,240, OpenCvCameraRotation.UPRIGHT);
-            }
-            @Override
-            public void onError(int errorCode)
-            {
-                //This will be called if the camera could not be opened
-            }
-        });
+
     }
 }
